Refractoring: extracted Discount class
